requirements.txt
streamlit
ultralytics
easyocr
opencv-python-headless
Pillow
numpy
torch
torchvision